Movement detection to find which frames contain just background
Split a clip into left half and right half. The detected movement
in left and right half for every frame by differencing the
frames. Then used zero crossing to find the frame # where the
cell moves from the left to the right frame. Now I "just"
need to calc the averages and merge the left and right together.