Complete rewrite of the training with separate train & val data
The training data comes now from a pool of avi clips to
have max variance.

The validation data comes from two other clips.

The frame generator has now simply a function for the
max number of clips supplied. These can be limited
for training as the list of clip/label pairs are shuffled.